LayoutTests/fast/shadow-dom/Element-interface-attachShadow.html

<!DOCTYPE html>
<html>
<head>
<title>Shadow DOM: Attaching a ShadowRoot</title>
<meta name="author" title="Ryosuke Niwa" href="mailto:rniwa@webkit.org">
<meta name="assert" content="Element.prototype.attachShadow should create an instance of ShadowRoot">
<link rel="help" href="https://w3c.github.io/webcomponents/spec/shadow/#widl-Element-attachShadow-ShadowRoot-ShadowRootInit-shadowRootInitDict">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel='stylesheet' href='../../resources/testharness.css'>
</head>
<body>
<div id="log"></div>
<script>

test(function () {
    assert_true('attachShadow' in Element.prototype, 'Element.prototype.attachShadow must exist');
    assert_equals(typeof(document.createElement('div').attachShadow), 'function', 'An instance of div must have attachShadow which is a function');
}, 'Check the existence of Element.attachShadow');

test(function () {
    assert_false('attachShadow' in Node.prototype, 'Node.prototype.attachShadow must not exist');
    assert_false('attachShadow' in CharacterData.prototype, 'CharacterData.prototype.attachShadow must not exist');
    assert_false('attachShadow' in Comment.prototype, 'Comment.prototype.attachShadow must not exist');
    assert_equals(typeof(document.createComment('').attachShadow), 'undefined', 'An instance of comment must not have attachShadow');
    assert_false('attachShadow' in Document.prototype, 'Document.prototype.attachShadow must not exist');
    assert_equals(typeof(document.attachShadow), 'undefined', 'An instance of document must not have attachShadow which is a function');
    assert_false('attachShadow' in DocumentFragment.prototype, 'DocumentFragment.prototype.attachShadow must not exist');
    assert_equals(typeof((new DOMParser()).parseFromString('', 'text/html').attachShadow), 'undefined', 'An instance of document must not have attachShadow which is a function');
    assert_false('attachShadow' in Text.prototype, 'Text.prototype.attachShadow must not exist');
    assert_equals(typeof(document.createTextNode('').attachShadow), 'undefined', 'An instance of text node must not have attachShadow');
}, 'Nodes other than Element should not have attachShadow');

test(function () {
    assert_throws({'name': 'TypeError'}, function () {
        document.createElement('div').attachShadow({})
    }, 'attachShadow must throw a TypeError when mode is omitted');

    assert_throws({'name': 'TypeError'}, function () {
        document.createElement('div').attachShadow({mode: true})
    }, 'attachShadow must throw a TypeError when mode is a boolean');

    assert_throws({'name': 'TypeError'}, function () {
        document.createElement('div').attachShadow({mode: 1})
    }, 'attachShadow must throw a TypeError when mode is 1');
}, 'Element.attachShadow must throw a TypeError if mode is not "open" or "closed"');

test(function () {
    assert_true(document.createElement('div').attachShadow({mode: "open"}) instanceof ShadowRoot,
        'attachShadow({mode: "open"}) should create an instance of ShadowRoot');
    assert_true(document.createElement('div').attachShadow({mode: "closed"}) instanceof ShadowRoot,
        'attachShadow({mode: "closed"}) should create an instance of ShadowRoot');
}, 'Element.attachShadow must create an instance of ShadowRoot');

test(function () {
    assert_throws({'name': 'InvalidStateError'}, function () {
        var div = document.createElement('div');
        div.attachShadow({mode: "open"});
        div.attachShadow({mode: "open"});
    }, 'Calling attachShadow({mode: "open"}) twice on the same element must throw');

    assert_throws({'name': 'InvalidStateError'}, function () {
        var div = document.createElement('div');
        div.attachShadow({mode: "closed"});
        div.attachShadow({mode: "closed"});
    }, 'Calling attachShadow({mode: "closed"}) twice on the same element must throw');

    assert_throws({'name': 'InvalidStateError'}, function () {
        var div = document.createElement('div');
        div.attachShadow({mode: "open"});
        div.attachShadow({mode: "closed"});
    }, 'Calling attachShadow({mode: "closed"}) after attachShadow({mode: "open"}) on the same element must throw');

    assert_throws({'name': 'InvalidStateError'}, function () {
        var div = document.createElement('div');
        div.attachShadow({mode: "closed"});
        div.attachShadow({mode: "open"});
    }, 'Calling attachShadow({mode: "open"}) after attachShadow({mode: "closed"}) on the same element must throw');
}, 'Element.attachShadow must throw a InvalidStateError if the context object already hosts a shadow tree');

var htmlElementNames = ['a', 'abbr', 'acronym', 'address', 'applet', 'area', 'article', 'aside', 'audio', 'b', 'base', 'basefont', 'bdi',
    'bdo', 'bgsound', 'big', 'blockquote', 'body', 'br', 'button', 'canvas', 'caption', 'center', 'cite', 'code', 'col', 'colgroup',
    'command', 'data', 'datalist', 'dd', 'del', 'details', 'dfn', 'dir', 'div', 'dl', 'dt', 'em', 'embed', 'fieldset', 'figcaption',
    'figure', 'font', 'footer', 'form', 'frame', 'frameset', 'h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'head', 'header', 'hgroup', 'hr',
    'html', 'i', 'iframe', 'image', 'img', 'input', 'ins', 'isindex', 'kbd', 'keygen', 'label', 'layer', 'legend', 'li', 'link', 'listing',
    'main', 'map', 'mark', 'marquee', 'menu', 'meta', 'meter', 'nav', 'nobr', 'noembed', 'noframes', 'nolayer', 'object', 'ol', 'optgroup',
    'option', 'output', 'p', 'param', 'picture', 'plaintext', 'pre', 'progress', 'q', 'rb', 'rp', 'rt', 'rtc', 'ruby', 's', 'samp', 'script',
    'section', 'select', 'slot', 'small', 'source', 'span', 'strike', 'strong', 'style', 'sub', 'summary', 'sup', 'table', 'tbody', 'td',
    'template', 'textarea', 'tfoot', 'th', 'thead', 'time', 'title', 'tr', 'track', 'tt', 'u', 'ul', 'var', 'video', 'wbr', 'xmp', 'noscript'];
var allowingAttachShadow = ["article", "aside", "blockquote", "body", "div", "footer", "h1", "h2", "h3", "h4", "h5", "h6", "header", "nav", "p", "section", "span"];

test(function () {
    for (var elementName of htmlElementNames) {
        if (allowingAttachShadow.indexOf(elementName) >= 0)
            continue;

        assert_throws({'name': 'NotSupportedError'}, function () {
            document.createElement(elementName).attachShadow({mode: "open"});
        }, 'Calling attachShadow({mode: "open"}) on ' + elementName + ' element must throw');

        assert_throws({'name': 'NotSupportedError'}, function () {
            document.createElement(elementName).attachShadow({mode: "closed"});
        }, 'Calling attachShadow({mode: "closed"}) on ' + elementName + ' element must throw');
    }
}, 'Element.attachShadow must throw a NotSupportedError on HTML elements whose local name is not one of ' + allowingAttachShadow.join(', '));

test(function () {
    for (var elementName of allowingAttachShadow) {
        assert_true(document.createElement(elementName).attachShadow({mode: "open"}) instanceof ShadowRoot,
            'Calling attachShadow({mode: "open"}) on ' + elementName + ' element must not throw');

        assert_true(document.createElement(elementName).attachShadow({mode: "closed"}) instanceof ShadowRoot,
            'Calling attachShadow({mode: "closed"}) on ' + elementName + ' element must not throw');
    }
}, 'Element.attachShadow must not throw a NotSupportedError on ' + allowingAttachShadow.join(', '));

test(function () {
    assert_throws({'name': 'NotSupportedError'}, function () {
        document.createElement('w3cfutureelement').attachShadow({mode: "open"});
    }, 'Calling attachShadow({mode: "open"}) on a HTML element whose local name is "w3cfutureelement" must throw');

    assert_throws({'name': 'NotSupportedError'}, function () {
        document.createElement('w3cfutureelement').attachShadow({mode: "closed"});
    }, 'Calling attachShadow({mode: "closed"}) on a HTML element whose local name is "w3cfutureelement" must throw');

}, 'Element.attachShadow must throw a NotSupportedError on a HTMLKnownElement which does not have a valid custom element');

test(function () {
    assert_true(document.createElement('custom-element').attachShadow({mode: "open"}) instanceof ShadowRoot,
        'Calling attachShadow({mode: "open"}) on a HTML element whose local name is "custom-element" element must not throw');

    assert_true(document.createElement('custom-element').attachShadow({mode: "closed"}) instanceof ShadowRoot,
        'Calling attachShadow({mode: "closed"}) on a HTML element whose local name is "custom-element" must not throw');

}, 'Element.attachShadow must throw a NotSupportedError on a HTMLKnownElement which has a valid custom element');

</script>
</body>
</html>
